moving to wizard approach for bulk download
