+++ config/routes.rb
@@ -90,10 +90,12 @@ Rails.application.routes.draw do
         scope :search do
           get 'facets', to: 'search#facets', as: :search_facets
           get 'facet_filters', to: 'search#facet_filters', as: :search_facet_filters
-          post 'auth_code', to: 'search#create_auth_code', as: :create_auth_code
-          get 'bulk_download', to: 'search#bulk_download', as: :search_bulk_download
-          get 'bulk_download_info', to: 'search#bulk_download_info', as: :search_bulk_download_info
           get '/', to: 'search#index', as: :search
+        end
+        scope :bulk_download do
+          post 'auth_code', to: 'bulk_download#auth_code', as: :bulk_download_auth_code
+          get 'summary', to: 'bulk_download#summary', as: :bulk_download_summary
+          get 'generate_curl_config', to: 'bulk_download#generate_curl_config', as: :bulk_download_generate_curl_config
         end
       end
     end
